Moved change strides into own file
